Refactored for cleaner code
Split some code up in more files for better readabillity
